% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_checkName_wcvp.R
\name{batch_checkName_wcvp}
\alias{batch_checkName_wcvp}
\title{In batch, use the WCVP database to check accepted names and update synonyms.}
\usage{
batch_checkName_wcvp(
  occ = NA,
  wcvp_names = NA,
  if_author_fails_try_without_combinations = TRUE,
  wcvp_selected_fields = "standard"
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard').}

\item{wcvp_names}{WCVP table, wcvp_names.csv file from http://sftp.kew.org/pub/data-repositories/WCVP/ If NA, automatically load the latest version of the database by the function ParsGBIF::get_wcvp(read_only_to_memory = TRUE)$wcvp_names.}

\item{if_author_fails_try_without_combinations}{option for partial verification of the authorship of the species. Remove the authors of combinations, in parentheses.}

\item{wcvp_selected_fields}{WCVP fields selected as return, 'standard' basic columns, 'all' all available columns.}
}
\value{
Data frame with WCVP fields
}
\description{
In batch, use the \href{https://powo.science.kew.org//}{World Checklist of Vascular Plants}
\href{http://sftp.kew.org/pub/data-repositories/WCVP/}{database}
\href{https://powo.science.kew.org/about-wcvp}{(about WCVP)} to check accepted names and update synonyms.
}
\details{
See help(checkName_wcvp) and \href{http://sftp.kew.org/pub/data-repositories/WCVP/}{about WCVP database}

#' @author Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
\examples{
# These examples take >10 seconds to run and require 'ParsGBIF::get_wcvp()'
\donttest{
# load package
library(ParsGBIF)

help(batch_checkName_wcvp)

wcvp_names <- get_wcvp(read_only_to_memory = TRUE)$wcvp_names

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file =  'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')

res_batch_checkName_wcvp <- batch_checkName_wcvp(occ = occ,
                                                 wcvp_names =  wcvp_names,
                                                 if_author_fails_try_without_combinations = TRUE,
                                                 wcvp_selected_fields = 'standard')

names(res_batch_checkName_wcvp)

head(res_batch_checkName_wcvp$wcvpSummary)

head(res_batch_checkName_wcvp$wcvpOccurrence)
}

}
\seealso{
\code{\link[ParsGBIF]{get_wcvp}}, \code{\link[ParsGBIF]{checkName_wcvp}}
}
