% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkName_wcvp.R
\name{checkName_wcvp}
\alias{checkName_wcvp}
\title{checkName_wcvp}
\usage{
checkName_wcvp(
  searchedName = "Hemistylus brasiliensis Wedd.",
  wcvp_names = wcvp_names,
  if_author_fails_try_without_combinations = TRUE
)
}
\arguments{
\item{searchedName}{scientific name, with or without author}

\item{wcvp_names}{WCVP table, wcvp_names.csv file from http://sftp.kew.org/pub/data-repositories/WCVP/}

\item{if_author_fails_try_without_combinations}{option for partial verification of the authorship of the species. Remove the authors of combinations, in parentheses}
}
\value{
wcvp_plant_name_id,
wcvp_taxon_rank,
wcvp_taxon_status,
wcvp_family,
wcvp_taxon_name,
wcvp_taxon_authors,
wcvp_accepted_plant_name_id,
wcvp_reviewed,
wcvp_searchedName,
wcvp_taxon_status_of_searchedName,
wcvp_plant_name_id_of_searchedName,
wcvp_taxon_authors_of_searchedName,
wcvp_verified_author,
wcvp_verified_speciesName,
wcvp_searchNotes.
}
\description{
Use the World Checklist of Vascular Plants (WCVP) database to
check accepted names and update synonyms
}
\details{
searchNotes values:
Accepted - When only one authorless scientific name is present in the list of TAXON_name with and TAXON_STATUS equal to "Accepted",
verified_speciesName = 100.
Accepted among homonyms - When more than one authorless scientific name is present in the TAXON_name list, but only one of the homonyms displays TAXON_STATUS equal to "Accepted",
verified_speciesName = number of matches/100.
Homonyms - When more than one authorless scientific name is present in the TAXON_name list and more than one, or none among the homonyms, display TAXON_STATUS equal to "Accepted",
verified_speciesName = number of matches/100.
Before searching for homonyms, there was a failure in trying to find the matching match between authorless scientific name in TAXON_name and author in TAXON_AUTHORS, in these cases verified_author equal to 0 (zero),
Not Found: When the authorless scientific name is not present in the TAXON_NAME LIST
Unplaced: o	When only one authorless scientific name is present in the list of TAXON_name with and TAXON_STATUS = "Unplaced"
Updated: When only one authorless scientific name is present in the list of TAXON_name and ACCEPTED_PLANT_NAME_ID are not empty (and ACCEPTED_PLANT_NAME_ID is different from the ID of the species consulted)
taxon_status_of_searchedName, plant_name_id_of_searchedName and taxon_authors_of_searchedName values:
When searchNotes equals "Updated" – The fields record the information of the scientific name originally consulted.
When searchNotes equals "Homonyms" - Fields record the information of homonymous synonyms separated by "|".
verified_author values:
When value equal to 100 – when there is matched match between authorless scientific name in TAXON_name and author in TAXON_AUTHORS.
When value equal to 50 – when there is combined correspondence between authorless scientific name in TAXON_name and author, without (combination), in TAXON_AUTHORS.
When value equal to 0 – regardless of the correspondence between authorless scientific name in TAXON_name, author is not present in TAXON_AUTHORS.
}
\examples{
utils::unzip(wcvp.zip, exdir = "c:/") 

wcvp_names <- read.table('c:/wcvp_names.csv', sep="|", header=TRUE, quote = "", fill=TRUE, encoding = "UTF-8")
  
checkName_wcvp(searchedName = 'Hemistylus brasiliensis Wedd.',
                  wcvp_names= wcvp_names,
                  if_author_fails_try_without_combinations = TRUE)
                  
wcvp <- get_wcvp(url_source = 'http://sftp.kew.org/pub/data-repositories/WCVP/',
                 path_results = 'c:/',
                 update = FALSE,
                 load_distribution = FALSE)
                  
checkName_wcvp(searchedName = 'Hemistylus brasiliensis Wedd.',
                  wcvp_names= wcvp$wcvp_names,
                  if_author_fails_try_without_combinations = TRUE)
}
\seealso{
\code{\link[utils]{unzip}}, \code{\link[unzip]{read.table}}
}
\author{
Pablo Hendrigo Alves de Melo

Nadia Bystriakova

Alexandre Monro
}
