% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EnumOccurrenceIssue.R
\docType{data}
\name{EnumOccurrenceIssue}
\alias{EnumOccurrenceIssue}
\title{Enumeration GBIF issue}
\format{
A data frame with 69 rows and 9 columns
}
\source{
\url{https://gbif.github.io/parsers/apidocs/org/gbif/api/vocabulary/OccurrenceIssue.html}
}
\usage{
data(EnumOccurrenceIssue)
}
\description{
An enumeration of validation rules for single occurrence records.
}
\details{
\describe{
\item{constant}{GBIF Constant}
\item{description}{Description}
\item{definition}{definition}
\item{type}{type}
\item{priority}{priority}
\item{score}{score}
\item{selection_score}{selection_score}
\item{reasoning}{reasoning}
\item{notes}{notes}
}
}
\keyword{datasets}
