% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{Export data to analyze or review}
\usage{
export_data(occ_digital_voucher_file = "", occ_digital_voucher = NA)
}
\arguments{
\item{occ_digital_voucher_file}{CSV fila result of function select_digital_voucher_and_sample_identification()$occ_digital_voucher_and_sample_identification}

\item{occ_digital_voucher}{data frame result of function select_digital_voucher_and_sample_identification()$occ_digital_voucher_and_sample_identification}
}
\value{
list with three data frames:
occ_in (data useful for spatial and taxonomic analysis),
occ_out_to_recover (data in need of spatial data revision or without identification) and
occ_dup (duplicates).
}
\description{
Separate records into data frames to analyze or review
}
\details{
...
}
\examples{
\donttest{
help(export_data)

}
}
\seealso{
\code{\link[ParsGBIF]{batch_checkName_wcvp}}, \code{\link[ParsGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
