% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lastNameRecordedBy.R
\name{get_lastNameRecordedBy}
\alias{get_lastNameRecordedBy}
\title{get_lastNameRecordedBy}
\usage{
get_lastNameRecordedBy(x)
}
\arguments{
\item{x}{recordedBy field}
}
\value{
last name of the main collector
}
\description{
Returns the last name of the main collector
}
\details{
Returns the last name of the main collector in recordedBy field
}
\examples{
help(get_lastNameRecordedBy)

get_lastNameRecordedBy('Melo, P.H.A & Monro, A.')

get_lastNameRecordedBy('Monro, A. & Melo, P.H.A')
}
\seealso{
\code{\link[ParsGBIF]{prepare_lastNameRecordedBy}}, \code{\link[ParsGBIF]{update_lastNameRecordedBy}}
}
\author{
Pablo Hendrigo Alves de Melo

Nadia Bystriakova

Alexandre Monro
}
