% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wcvp.R
\name{get_wcvp}
\alias{get_wcvp}
\title{get_wcvp}
\usage{
get_wcvp(
  url_source = "http://sftp.kew.org/pub/data-repositories/WCVP/",
  read_only_to_memory = FALSE,
  path_results = "C:/ParsGBIF",
  update = FALSE,
  load_distribution = FALSE
)
}
\arguments{
\item{url_source}{http://sftp.kew.org/pub/data-repositories/WCVP/}

\item{read_only_to_memory}{TRUE to in-memory read-only, not writing a copy to local disk}

\item{path_results}{download destination folder, if read_only_to_memory FALSE}

\item{update}{TRUE to update and load files, FALSE to keep local version and load files, if read_only_to_memory FALSE}

\item{load_distribution}{TRUE to load file with geographical distribution of species, if read_only_to_memory FALSE}
}
\value{
README_WCVP.xlsx,
wcvp_distribution.csv,
wcvp_names.csv
}
\description{
Download World Checklist of Vascular Plants (WCVP) database
}
\details{
http://sftp.kew.org/pub/data-repositories/WCVP/ This is the public SFTP (Secure File Transfer Protocol) site of the Royal Botanic Gardens, Kew. This space contains data resources publicly accessible to the user `anonymous'.  No password required for access. Use of data made available via this site may be subject to legal and licensing restrictions. The README in the top-level directory for each data resource provides specific information about its terms of use.
}
\examples{
# load package
library(ParsGBIF)

help(get_wcvp)

# Download wcvp database to local disk
path_data <- tempdir() # you can change this folder

wcvp <- get_wcvp(url_source = 'http://sftp.kew.org/pub/data-repositories/WCVP/',
                 read_only_to_memory = FALSE,
                 path_results = path_data,
                 update = FALSE,
                 load_distribution = TRUE)

names(wcvp)

head(wcvp$wcvp_names)
colnames(wcvp$wcvp_names)

head(wcvp$wcvp_distribution)
colnames(wcvp$wcvp_distribution)
}
\seealso{
\code{\link[ParsGBIF]{checkName_wcvp}}, \code{\link[ParsGBIF]{standardize_scientificName}}
}
\author{
Pablo Hendrigo Alves de Melo

Nadia Bystriakova

Alexandre Monro
}
