% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepere_lastNameRecordedBy.R
\name{prepere_lastNameRecordedBy}
\alias{prepere_lastNameRecordedBy}
\title{prepere_lastNameRecordedBy}
\usage{
prepere_lastNameRecordedBy(occ = NA, collectorDictionary_file = NA)
}
\arguments{
\item{occ}{GBIF occurrence file with selected columns}

\item{collectorDictionary_file}{Collector's dictionary file}
}
\value{
Ctrl_nameRecordedBy_Standard,
Ctrl_recordedBy,
Ctrl_notes,
collectorDictionary,
Ctrl_update,
collectorName,
Ctrl_fullName,
Ctrl_fullNameII,
CVStarrVirtualHerbarium_PersonDetails
}
\description{
Returns the last name of the primary collector.
If recordedBy is present in the collector's dictionary, it returns the checked name, if not, it returns the last name of the main collector, extracted from the recordedBy field.
}
\details{
....
}
\examples{
collectorDictionary <- readr::read_csv('Collectors Dictionary.csv', 
                                       locale = readr::locale(encoding = "UTF-8"))
 
collectorsDictionaryFromDataset <- prepere_lastNameRecordedBy(occ=occ,
                                                              collectorDictionary=collectorDictionary)

}
\seealso{
\code{\link[readr]{read_csv}}, \code{\link[unzip]{read.table}}
}
\author{
Pablo Hendrigo Alves de Melo

Nadia Bystriakova

Alexandre Monro
}
