% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/select_digital_voucher_and_sample_identification.R
\name{select_digital_voucher_and_sample_identification}
\alias{select_digital_voucher_and_sample_identification}
\title{select_digital_voucher_and_sample_identification}
\usage{
select_digital_voucher_and_sample_identification(
  occ = NA,
  occ_gbif_issue = NA,
  occ_checkName_wcvp = NA,
  occ_collectorsDictionary = NA,
  enumOccurrenceIssue = NA
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{occ_gbif_issue}{= result of function extract_gbif_issue()$occ_gbif_issue}

\item{occ_checkName_wcvp}{= result of function batch_checkName_wcvp()$occ_checkName_wcvp}

\item{occ_collectorsDictionary}{= result of function update_collectorsDictionary()$occ_collectorsDictionary}

\item{enumOccurrenceIssue}{An enumeration of validation rules for single occurrence records by GBIF file, if NA, will be used, data(EnumOccurrenceIssue)}
}
\value{
matchStatusDuplicates - "matched", "unmatched: no recordedBy and no recordNumber", "unmatched: no recordNumber" or "unmatched: no recordedBy"
numberTaxonNamesSample -  count of the different accepted scientific names, identified up to or below the specific level, listed in the sample duplicates, or Zero, if there is no identification, equal to or below the specific level, for the sample.
sampleTaxonName - TAXON_name accepted and identified up to or below the specific level selected for the sample.
sampleIdentificationStatus - 'Identified', 'divergent identifications', or 'unidentified'
}
\description{
Extract gbif issue
}
\details{
To group duplicates: 1) If the key to group duplicates is incomplete: Sample duplicates cannot be grouped due to missing collector information and/or collection number. Each record is considered a sample, with no duplicates. Select a voucher for each sample.
Or, 2) If the key to group duplicates is complete: Group duplicates. Select a voucher, with the highest information score, among the duplicates in the sample.
How to calculate the information score?
moreInformativeRecord = sum of verbatim quality + quality of geospatial information. verbatim quality = sum of the number of flags with verbatim quality equal to TRUE. quality of geospatial information = If there is a geospatial issue, consider the one with the highest priority, with the highest score.
To select sample taxonomic identification: 1) If the key for grouping duplicates is complete: select the accepted TAXON_NAME, identified up to or below the specific level, most frequent among the duplicates in the sample. If tie between frequency of accepted TAXON_NAME, identified up to or below the specific level: select the first accepted TAXON_NAME, identified up to or below the specific level, in alphabetical order. If there is no identification, at or below the specific level, for the sample: Indicate as unidentified sample.
Or, 2) If the key for grouping duplicates is incomplete: select TAXON_NAME, if accepted and identified up to or below the specified level. If there is no identification, at or below the specific level, for the sample: Indicate as unidentified sample.
}
\examples{
\donttest{
help(select_digital_voucher_and_sample_identification)

head(occ)
head(res_gbif_issue$occ_gbif_issue)
head(res_checkName_wcvp$occ_checkName_wcvp)
head(res_collectorsDictionary$occ_collectorsDictionary)
res_digital_voucher_and_sample_identification <- select_digital_voucher_and_sample_identification(occ = occ,
                                                                                                  occ_gbif_issue = res_gbif_issue$occ_gbif_issue,
                                                                                                  occ_checkName_wcvp = res_checkName_wcvp$occ_checkName_wcvp,
                                                                                                  occ_collectorsDictionary = res_collectorsDictionary$occ_collectorsDictionary,
                                                                                                  enumOccurrenceIssue = EnumOccurrenceIssue)

names(res_digital_voucher_and_sample_identification)

head(res_digital_voucher_and_sample_identification$occ_digital_voucher_and_sample_identification)
colnames(res_digital_voucher_and_sample_identification$occ_digital_voucher_and_sample_identification)

head(res_digital_voucher_and_sample_identification$occ_join_results)
colnames(res_digital_voucher_and_sample_identification$occ_join_results)
}
}
\seealso{
\code{\link[ParsGBIF]{batch_checkName_wcvp}}, \code{\link[ParsGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
