% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_gbif_fields.R
\name{select_gbif_fields}
\alias{select_gbif_fields}
\title{select_gbif_fields}
\usage{
select_gbif_fields(columns = "standard")
}
\arguments{
\item{columns}{'standard' basic columns about what, when, where, and who collected, 'all' all available columns or list column names}
}
\value{
list of the columns names
}
\description{
Select columns in GBIF occurrence data
}
\details{
standard:
c(
'bibliographicCitation',
'language',
'institutionCode',
'collectionCode',
'datasetName',
'basisOfRecord',
'informationWithheld',
'dataGeneralizations'/,
'occurrenceID', # occ_search(occurrenceId='BRA:UNEMAT:HPAN:6089')
'catalogNumber',
'recordNumber',
'recordedBy',
'georeferenceVerificationStatus',
'occurrenceStatus',
'eventDate',
'year',
'month',
'day',
'habitat',
'fieldNotes',
'eventRemarks',
'locationID',
'higherGeography',
'islandGroup',
'island',
'countryCode',
'stateProvince',
'county',
'municipality',
'locality',
'verbatimLocality',
'locationRemarks',
'decimalLatitude',
'decimalLongitude',
'verbatimCoordinateSystem',
'verbatimIdentification',
'identificationQualifier',
'typeStatus',
'identifiedBy',
'dateIdentified',
'scientificName',
'family',
'taxonRank',
'nomenclaturalCode',
'taxonomicStatus',
'issue',
'mediaType',
'hasCoordinate',
'hasGeospatialIssues',
'verbatimScientificName',
'level0Name',
'level1Name',
'level2Name',
'level3Name')

'all':
c(
'gbifID',
'abstract',
'accessRights',
'accrualMethod',
'accrualPeriodicity',
'accrualPolicy',
'alternative',
'audience',
'available',
'bibliographicCitation',
'conformsTo',
'contributor',
'coverage',
'created',
'creator',
'date',
'dateAccepted',
'dateCopyrighted',
'dateSubmitted',
'description',
'educationLevel',
'extent',
'format',
'hasFormat',
'hasPart',
'hasVersion',
'identifier',
'instructionalMethod',
'isFormatOf',
'isPartOf',
'isReferencedBy',
'isReplacedBy',
'isRequiredBy',
'isVersionOf',
'issued',
'language',
'license',
'mediator',
'medium',
'modified',
'provenance',
'publisher',
'references',
'relation',
'replaces',
'requires',
'rights',
'rightsHolder',
'source',
'spatial',
'subject',
'tableOfContents',
'temporal',
'title',
'type',
'valid',
'institutionID',
'collectionID',
'datasetID',
'institutionCode',
'collectionCode',
'datasetName',
'ownerInstitutionCode',
'basisOfRecord',
'informationWithheld',
'dataGeneralizations',
'dynamicProperties',
'occurrenceID',
'catalogNumber',
'recordNumber',
'recordedBy',
'recordedByID',
'individualCount',
'organismQuantity',
'organismQuantityType',
'sex',
'lifeStage',
'reproductiveCondition',
'behavior',
'establishmentMeans',
'degreeOfEstablishment',
'pathway',
'georeferenceVerificationStatus',
'occurrenceStatus',
'preparations',
'disposition',
'associatedOccurrences',
'associatedReferences',
'associatedSequences',
'associatedTaxa',
'otherCatalogNumbers',
'occurrenceRemarks',
'organismID',
'organismName',
'organismScope',
'associatedOrganisms',
'previousIdentifications',
'organismRemarks',
'materialSampleID',
'eventID',
'parentEventID',
'fieldNumber',
'eventDate',
'eventTime',
'startDayOfYear',
'endDayOfYear',
'year',
'month',
'day',
'verbatimEventDate',
'habitat',
'samplingProtocol',
'sampleSizeValue',
'sampleSizeUnit',
'samplingEffort',
'fieldNotes',
'eventRemarks',
'locationID',
'higherGeographyID',
'higherGeography',
'continent',
'waterBody',
'islandGroup',
'island',
'countryCode',
'stateProvince',
'county',
'municipality',
'locality',
'verbatimLocality',
'verbatimElevation',
'verticalDatum',
'verbatimDepth',
'minimumDistanceAboveSurfaceInMeters',
'maximumDistanceAboveSurfaceInMeters',
'locationAccordingTo',
'locationRemarks',
'decimalLatitude',
'decimalLongitude',
'coordinateUncertaintyInMeters',
'coordinatePrecision',
'pointRadiusSpatialFit',
'verbatimCoordinateSystem',
'verbatimSRS',
'footprintWKT',
'footprintSRS',
'footprintSpatialFit',
'georeferencedBy',
'georeferencedDate',
'georeferenceProtocol',
'georeferenceSources',
'georeferenceRemarks',
'geologicalContextID',
'earliestEonOrLowestEonothem',
'latestEonOrHighestEonothem',
'earliestEraOrLowestErathem',
'latestEraOrHighestErathem',
'earliestPeriodOrLowestSystem',
'latestPeriodOrHighestSystem',
'earliestEpochOrLowestSeries',
'latestEpochOrHighestSeries',
'earliestAgeOrLowestStage',
'latestAgeOrHighestStage',
'lowestBiostratigraphicZone',
'highestBiostratigraphicZone',
'lithostratigraphicTerms',
'group',
'formation',
'member',
'bed',
'identificationID',
'verbatimIdentification',
'identificationQualifier',
'typeStatus',
'identifiedBy',
'identifiedByID',
'dateIdentified',
'identificationReferences',
'identificationVerificationStatus',
'identificationRemarks',
'taxonID',
'scientificNameID',
'acceptedNameUsageID',
'parentNameUsageID',
'originalNameUsageID',
'nameAccordingToID',
'namePublishedInID',
'taxonConceptID',
'scientificName',
'acceptedNameUsage',
'parentNameUsage',
'originalNameUsage',
'nameAccordingTo',
'namePublishedIn',
'namePublishedInYear',
'higherClassification',
'kingdom',
'phylum',
'class',
'order',
'family',
'subfamily',
'genus',
'genericName',
'subgenus',
'infragenericEpithet',
'specificEpithet',
'infraspecificEpithet',
'cultivarEpithet',
'taxonRank',
'verbatimTaxonRank',
'vernacularName',
'nomenclaturalCode',
'taxonomicStatus',
'nomenclaturalStatus',
'taxonRemarks',
'datasetKey',
'publishingCountry',
'lastInterpreted',
'elevation',
'elevationAccuracy',
'depth',
'depthAccuracy',
'distanceAboveSurface',
'distanceAboveSurfaceAccuracy',
'issue',
'mediaType',
'hasCoordinate',
'hasGeospatialIssues',
'taxonKey',
'acceptedTaxonKey',
'kingdomKey',
'phylumKey',
'classKey',
'orderKey',
'familyKey',
'genusKey',
'subgenusKey',
'speciesKey',
'species',
'acceptedScientificName',
'verbatimScientificName',
'typifiedName',
'protocol',
'lastParsed',
'lastCrawled',
'repatriated',
'relativeOrganismQuantity',
'level0Gid',
'level0Name',
'level1Gid',
'level1Name',
'level2Gid',
'level2Name',
'level3Gid',
'level3Name',
'iucnRedListCategory')
}
\examples{
# select_gbif_fields()

help(select_gbif_fields)

col_sel <- select_gbif_fields(columns = 'all')

col_sel <- select_gbif_fields(columns = 'standard')

}
\seealso{
\code{\link[ParsGBIF]{extract_gbif_issue}}, \code{\link[ParsGBIF]{prepare_gbif_occurrence_data}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
