% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_checkName_wcvp.R
\name{batch_checkName_wcvp}
\alias{batch_checkName_wcvp}
\title{batch_checkName_wcvp}
\usage{
batch_checkName_wcvp(
  occ = NA,
  wcvp_names = wcvp_names,
  if_author_fails_try_without_combinations = TRUE,
  wcvp_selected_fields = "standard"
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{wcvp_names}{WCVP table, wcvp_names.csv file from http://sftp.kew.org/pub/data-repositories/WCVP/}

\item{if_author_fails_try_without_combinations}{option for partial verification of the authorship of the species. Remove the authors of combinations, in parentheses}

\item{wcvp_selected_fields}{WCVP fields selected as return, 'standard' basic columns, 'all' all available columns}
}
\value{
occ_checkName_wcvp:
wcvp_plant_name_id,
wcvp_taxon_rank,
wcvp_taxon_status,
wcvp_family,
wcvp_taxon_name,
wcvp_taxon_authors,
wcvp_accepted_plant_name_id,
wcvp_reviewed,
wcvp_searchedName,
wcvp_taxon_status_of_searchedName,
wcvp_plant_name_id_of_searchedName,
wcvp_taxon_authors_of_searchedName,
wcvp_verified_author,
wcvp_verified_speciesName,
wcvp_searchNotes
summary
}
\description{
Use the World Checklist of Vascular Plants (WCVP) database
to check accepted names and update synonyms, in batch
}
\details{
See help(checkName_wcvp) and https://powo.science.kew.org/about-wcvp
}
\examples{
# load package
library(ParsGBIF)

help(batch_checkName_wcvp)

wcvp_names <- get_wcvp(read_only_to_memory = TRUE)$wcvp_names

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file =  'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')

res_batch_checkName_wcvp <- batch_checkName_wcvp(occ = occ,
                                                 wcvp_names =  wcvp_names,
                                                 if_author_fails_try_without_combinations = TRUE,
                                                 wcvp_selected_fields = 'standard')

names(res_batch_checkName_wcvp)

head(res_batch_checkName_wcvp$wcvpSummary)

head(res_batch_checkName_wcvp$wcvpOccurrence)
}
\seealso{
\code{\link[ParsGBIF]{get_wcvp}}, \code{\link[ParsGBIF]{checkName_wcvp}}
}
\author{
Pablo Hendrigo Alves de Melo
Nadia Bystriakova
Alexandre Monro
}
