% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gbif_issue.R
\name{extract_gbif_issue}
\alias{extract_gbif_issue}
\title{extract_gbif_issue}
\usage{
extract_gbif_issue(occ = NA, enumOccurrenceIssue = NA)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{enumOccurrenceIssue}{An enumeration of validation rules for single occurrence records by GBIF file, if NA, will be used, data(EnumOccurrenceIssue)}
}
\value{
occ_gbif_issue,
summary
}
\description{
Extract gbif issue
}
\details{
https://gbif.github.io/parsers/apidocs/org/gbif/api/vocabulary/OccurrenceIssue.html
}
\examples{
# extract_gbif_issue()

help(extract_gbif_issue)

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file =  'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')

data(EnumOccurrenceIssue)
head(EnumOccurrenceIssue)
colnames(EnumOccurrenceIssue)

occ_gbif_issue <- extract_gbif_issue(occ = occ,
                                     enumOccurrenceIssue = NA)

names(occ_gbif_issue)

head(occ_gbif_issue$issueGBIFSummary)

colnames(occ_gbif_issue$issueGBIFOccurrence)
head(occ_gbif_issue$issueGBIFOccurrence)
colnames(occ_gbif_issue$issueGBIFOccurrence)
}
\seealso{
\code{\link[ParsGBIF]{prepare_gbif_occurrence_data}}, \code{\link[ParsGBIF]{select_gbif_fields}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
