% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_gbif_issue.R
\name{extract_gbif_issue}
\alias{extract_gbif_issue}
\title{extract_gbif_issue}
\usage{
extract_gbif_issue(
  occ = NA,
  enumOccurrenceIssue_file = "C:/ParsGBIF/data/EnumOccurrenceIssue.csv"
)
}
\arguments{
\item{occ}{GBIF occurrence file with selected columns}

\item{enumOccurrenceIssue_file}{An enumeration of validation rules for single occurrence records by GBIF file}
}
\value{
issueGBIFSummary,
issueGBIFOccurrence
}
\description{
Extract gbif issue
}
\details{
https://gbif.github.io/parsers/apidocs/org/gbif/api/vocabulary/OccurrenceIssue.html
}
\examples{
extract_gbif_issue <- function(occ = occ,
                               enumOccurrenceIssue_file = "C:/ParsGBIF/data/EnumOccurrenceIssue.csv")
}
\seealso{
\code{\link[utils]{unzip}}, \code{\link[unzip]{read.table}}
}
\author{
Pablo Hendrigo Alves de Melo

Nadia Bystriakova

Alexandre Monro
}
