% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gbif_occurrence_data.R
\name{prepare_gbif_occurrence_data}
\alias{prepare_gbif_occurrence_data}
\title{Prepare occurrence data from GBIF to use in package}
\usage{
prepare_gbif_occurrence_data(gbif_occurrece_file = "", columns = "standard")
}
\arguments{
\item{gbif_occurrece_file}{the name of the file from which the GBIF herbarium specimen occurrence
data is to be read}

\item{columns}{Character vector of strings to indicate column names of the GBIF occurrence file.
Use 'standard' to select basic columns for use in the package, 'all' to select all available columns.
The default is 'standard'}
}
\value{
data.frame with renamed selected fields with prefix "Ctrl_"
}
\description{
Prepare GBIF herbarium specimen occurrence data for use in the package
}
\details{
Select data fields and rename field names prefixed with "Ctrl_"
}
\examples{
\donttest{

help(prepare_gbif_occurrence_data)

occ_file <- 'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/
occurrence.txt'

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file = occ_file,
                                    columns = 'standard')

colnames(occ)

head(occ)
}
}
\seealso{
\code{\link[ParsGBIF]{select_gbif_fields}}, \code{\link[ParsGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
