% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_gbif_occurrence_data.R
\name{prepare_gbif_occurrence_data}
\alias{prepare_gbif_occurrence_data}
\title{prepare_gbif_occurrence_data}
\usage{
prepare_gbif_occurrence_data(
  gbif_occurrece_file =
    "https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt",
  columns = "standard"
)
}
\arguments{
\item{gbif_occurrece_file}{the name of the file which the data are to be read from.
Either a path to a file, a connection, or literal data (either a single string or a raw vector)}

\item{columns}{'standard' by default. See select_gbif_fields() function.
*'standard' basic columns about what, when, where, and who collected,
*'all' all available columns
*or list column names}
}
\value{
data.frame with fields selected by the select_gbif_fields function and with "Ctrl_" at the beginning of each field name
}
\description{
Prepare occurrence data from GBIF to use in package
}
\details{
Prepare occurrence data from GBIF to use in package.
Select the data fields to be used.
Add "Ctrl_" at the beginning of each field name
}
\examples{
\donttest{
# prepare_gbif_occurrence_data()

help(prepare_gbif_occurrence_data)

occ <- prepare_gbif_occurrence_data(gbif_occurrece_file =
'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')

colnames(occ)

head(occ)
}
}
\seealso{
\code{\link[ParsGBIF]{select_gbif_fields}}, \code{\link[ParsGBIF]{extract_gbif_issue}}
}
\author{
Pablo Hendrigo Alves de Melo,
Nadia Bystriakova &
Alexandre Monro
}
