% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepere_collectorsDictionary.R
\name{prepere_collectorsDictionary}
\alias{prepere_collectorsDictionary}
\title{prepere_collectorsDictionary}
\usage{
prepere_collectorsDictionary(
  occ = NA,
  collectorDictionary_file =
    "https://raw.githubusercontent.com/pablopains/ParsGBIF/main/collectorDictionary/CollectorsDictionary.csv"
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{collectorDictionary_file}{Collector dictionary file - point to a file on your local disk or upload via git at https://raw.githubusercontent.com/pablopains/ParsGBIF/main/collectorDictionary/CollectorsDictionary.csv.}
}
\value{
Ctrl_nameRecordedBy_Standard,
Ctrl_recordedBy,
Ctrl_notes,
collectorDictionary,
Ctrl_update,
collectorName,
Ctrl_fullName,
Ctrl_fullNameII,
CVStarrVirtualHerbarium_PersonDetails
}
\description{
Returns the list with the last name of the main collector associated with the unique key recordedBy.
}
\details{
If recordedBy is present in the collector's dictionary, it returns the checked name, if not, it returns the last name of the main collector, extracted from the recordedBy field.
If recordedBy is present in the collector's dictionary, returns the main collector's last name associated with the single recordedBy key,
otherwise, returns the main collector's last name, extracted from the recordedBy field.
It is recommended to curate the main collector's surname, automatically extracted from the recordedBy field.
The objective is to standardize the last name of the main collector.
That the primary botanical collector of a sample is always recognized by the same last name, standardized in capital letters and non-ascii characters replaced
}
\examples{
help(prepere_collectorsDictionary)

occ <- prepere_gbif_occurrence_data(gbif_occurrece_file =  'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/dataGBIF/Achatocarpaceae/occurrence.txt',
                                    columns = 'standard')

collectorsDictionaryFromDataset <- prepere_collectorsDictionary(occ=occ,
                                                                collectorDictionary_file =  'https://raw.githubusercontent.com/pablopains/ParsGBIF/main/collectorDictionary/CollectorsDictionary.csv')

colnames(collectorsDictionaryFromDataset)
head(collectorsDictionaryFromDataset)

write.csv(collectorsDictionaryFromDataset,
          'collectorsDictionaryFromDataset.csv',
          row.names = FALSE,
          fileEncoding = "UTF-8",
          na = "")
}
\seealso{
\code{\link[ParsGBIF]{select_gbif_fields}}, \code{\link[ParsGBIF]{update_collectorsDictionary}},
\code{\link[textclean]{replace_non_ascii}
}
\author{
Pablo Hendrigo Alves de Melo

Nadia Bystriakova

Alexandre Monro
}
