% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepere_lastNameRecordedBy.R
\name{prepere_lastNameRecordedBy}
\alias{prepere_lastNameRecordedBy}
\title{prepere_lastNameRecordedBy}
\usage{
prepere_lastNameRecordedBy(
  occ = NA,
  collectorDictionary_url =
    "https://docs.google.com/spreadsheets/d/15Ngrr4hbJnq_SsTycLJ6z15oCLPRyV2gFhQ3D1zXzuk/edit?usp=share_link",
  collectorDictionary = NA
)
}
\arguments{
\item{occ}{GBIF occurrence table with selected columns as select_gbif_fields(columns = 'standard')}

\item{collectorDictionary_url}{Collector's dictionary URL curated by the ParsGBIF team 'https://docs.google.com/spreadsheets/d/15Ngrr4hbJnq_SsTycLJ6z15oCLPRyV2gFhQ3D1zXzuk/edit?usp=share_link'}

\item{collectorDictionary}{Collector's dictionary in data.frame format with the fields: Ctrl_nameRecordedBy_Standard, Ctrl_recordedBy, Ctrl_notes, collectorDictionary, Ctrl_update, collectorName, Ctrl_fullName, Ctrl_fullNameII, CVStarrVirtualHerbarium_PersonDetails.
if you prefer, download a CollectorsDictionary.csv template from https://drive.google.com/file/d/1sYh1s39Ee3JgMQp2iyePOTdCB9gbaotW/view?usp=share_link}
}
\value{
Ctrl_nameRecordedBy_Standard,
Ctrl_recordedBy,
Ctrl_notes,
collectorDictionary,
Ctrl_update,
collectorName,
Ctrl_fullName,
Ctrl_fullNameII,
CVStarrVirtualHerbarium_PersonDetails
}
\description{
Returns the last name of the primary collector.
If recordedBy is present in the collector's dictionary, it returns the checked name, if not, it returns the last name of the main collector, extracted from the recordedBy field.
}
\details{
....
}
\examples{
collectorsDictionaryFromDataset <- prepere_lastNameRecordedBy(occ=occ,
                                                              collectorDictionary_url='https://docs.google.com/spreadsheets/d/15Ngrr4hbJnq_SsTycLJ6z15oCLPRyV2gFhQ3D1zXzuk/edit?usp=share_link')

#' collectorsDictionaryFromDataset <- prepere_lastNameRecordedBy(occ=occ)

}
\seealso{
\code{\link[ParsGBIF]{select_gbif_fields}}, \code{\link[ParsGBIF]{update_lastNameRecordedBy}}
}
\author{
Pablo Hendrigo Alves de Melo

Nadia Bystriakova

Alexandre Monro
}
