% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/select_digital_voucher_and_sample_identification.R
\name{select_digital_voucher_and_sample_identification}
\alias{select_digital_voucher_and_sample_identification}
\title{select_digital_voucher_and_sample_identification}
\usage{
select_digital_voucher_and_sample_identification(
  occurrence_collectorsDictionary_file = NA,
  issueGBIFOccurrence_file = NA,
  wcvp_occurence_file = NA,
  enumOccurrenceIssue_file = "C:/ParsGBIF/data/EnumOccurrenceIssue.csv"
)
}
\arguments{
\item{occurrence_collectorsDictionary_file}{result of the update_lastNameRecordedBy function}

\item{issueGBIFOccurrence_file}{result of the extract_gbif_issue function}

\item{wcvp_occurence_file}{result of the checkName_wcvp function}

\item{enumOccurrenceIssue_file}{An enumeration of validation rules for single occurrence records by GBIF file}
}
\value{
matchStatusDuplicates - "matched", "unmatched: no recordedBy and no recordNumber", "unmatched: no recordNumber" or "unmatched: no recordedBy"
numberTaxonNamesSample -  count of the different accepted scientific names, identified up to or below the specific level, listed in the sample duplicates, or Zero, if there is no identification, equal to or below the specific level, for the sample.
sampleTaxonName - TAXON_name accepted and identified up to or below the specific level selected for the sample.
sampleIdentificationStatus - 'Identified', 'divergent identifications', or 'unidentified'
}
\description{
Extract gbif issue
}
\details{
To group duplicates: 1) If the key to group duplicates is incomplete: Sample duplicates cannot be grouped due to missing collector information and/or collection number. Each record is considered a sample, with no duplicates. Select a voucher for each sample.
Or, 2) If the key to group duplicates is complete: Group duplicates. Select a voucher, with the highest information score, among the duplicates in the sample.
How to calculate the information score?
moreInformativeRecord = sum of verbatim quality + quality of geospatial information. verbatim quality = sum of the number of flags with verbatim quality equal to TRUE. quality of geospatial information = If there is a geospatial issue, consider the one with the highest priority, with the highest score.
To select sample taxonomic identification: 1) If the key for grouping duplicates is complete: select the accepted TAXON_NAME, identified up to or below the specific level, most frequent among the duplicates in the sample. If tie between frequency of accepted TAXON_NAME, identified up to or below the specific level: select the first accepted TAXON_NAME, identified up to or below the specific level, in alphabetical order. If there is no identification, at or below the specific level, for the sample: Indicate as unidentified sample.
Or, 2) If the key for grouping duplicates is incomplete: select TAXON_NAME, if accepted and identified up to or below the specified level. If there is no identification, at or below the specific level, for the sample: Indicate as unidentified sample.
}
\examples{
occ <- select_digital_voucher_and_sample_identification(occurrence_collectorsDictionary_file = occurrence_collectorsDictionary_file,
                                                        issueGBIFOccurrence_file = issueGBIFOccurrence_file,
                                                        wcvp_occurence_file = wcvp_result_file,
                                                        enumOccurrenceIssue_file = enumOccurrenceIssue_file)
}
\seealso{
\code{\link[utils]{unzip}}, \code{\link[unzip]{read.table}}
}
\author{
Pablo Hendrigo Alves de Melo

Nadia Bystriakova

Alexandre Monro
}
