% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_lastNameRecordedBy.R
\encoding{UTF-8}
\name{update_lastNameRecordedBy}
\alias{update_lastNameRecordedBy}
\title{update_lastNameRecordedBy}
\usage{
update_lastNameRecordedBy(
  occ = NA,
  collectorDictionary_checked_file = NA,
  collectorDictionary_file = NA
)
}
\arguments{
\item{occ}{GBIF occurrence file with selected columns}

\item{collectorDictionary_checked_file}{Verified collector's dictionary file}

\item{collectorDictionary_file}{Collector's dictionary file}
}
\value{
collectorsDictionary_new,
recordNumber_Standard,
collectorsDictionary_summary
}
\description{
Include recordedByStandardized field with verified main collector's last name.
Include recordNumber_Standard field with only numbers from recordNumber.
Create a key to group duplicates in the key_family_recordedBy_recordNumber field, composed of the fields: family + recordedByStandardized + recordNumber_Standard.
}
\details{
....
}
\examples{
collectorDictionary_checked <- readr::read_csv('collectorDictionary_checked.csv', 
                                       locale = readr::locale(encoding = "UTF-8"))

collectorDictionary <- readr::read_csv('Collectors Dictionary.csv', 
                                       locale = readr::locale(encoding = "UTF-8"))
 
collectorsDictionaryFromDataset <- prepere_lastNameRecordedBy(occ=occ,
                                                              collectorDictionary=collectorDictionary)

}
\seealso{
\code{\link[readr]{read_csv}}, \code{\link[unzip]{read.table}}
}
\author{
Pablo Hendrigo Alves de Melo

Nadia Bystriakova

Alexandre Monro
}
